Added first insert version
The style and errors are not handled yet, also modalitats and serveis are set as null for the moment. It is just a first version in order to check the functionallity.
Added connection to insert a new Espai on DataAccess and added text fields, text areas, combobox, etc. On the insert panel of the mainform and a button to insert that calls DataAcces.InsertEspai().